#134 refactoring SqlBuilderTests
