+++ restservice/src/test/java/org/itevents/dao/mybatis/builder/FilterSqlBuilderTest.java
@@ -1,14 +1,8 @@
 package org.itevents.dao.mybatis.builder;
 
 import org.itevents.model.Filter;
-import org.itevents.model.Technology;
 import org.itevents.test_utils.BuilderUtil;
-import org.junit.Before;
 import org.junit.Test;
-
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
 
 import static org.junit.Assert.assertEquals;
 
@@ -17,119 +11,13 @@ import static org.junit.Assert.assertEquals;
  */
 public class FilterSqlBuilderTest {
 
-    private Filter parameter;
-
-    @Before
-    public void setup() {
-        parameter = new Filter();
-    }
-
-    @Test
-    public void shouldBuildSqlQueryWithEmptyParametersAndPagination() throws Exception {
-        String expectedSql = "SELECT * FROM event e WHERE (e.event_date > NOW()) " +
-                "ORDER BY event_date LIMIT #{limit} OFFSET #{offset}";
-        String returnedSql = new EventSqlBuilder().getFilteredEvents(parameter).replace('\n', ' ');
-        assertEquals(expectedSql, returnedSql);
-    }
-
-    @Test
-    public void shouldBuildSqlQueryByDateCityIdEventIdAndPagination() throws Exception {
-        parameter.setCity(BuilderUtil.buildCityKyiv());
-        List<Technology> testTechnologies = new ArrayList<>();
-        testTechnologies.add(BuilderUtil.buildTechnologyJava());
-        parameter.setTechnologies(testTechnologies);
-
-        String expectedSql = "SELECT * FROM event e JOIN event_technology et ON et.technology_id=-1 " +
-                "WHERE (e.event_date > NOW() AND city_id = #{city.id} AND e.id=et.event_id) " +
-                "ORDER BY event_date LIMIT #{limit} OFFSET #{offset}";
-
-        String returnedSql = new EventSqlBuilder().getFilteredEvents(parameter).replace('\n', ' ');
-        assertEquals(expectedSql, returnedSql);
-    }
-
-    @Test
-    public void shouldBuildSqlQueryByDateCityIdPriceAndPagination() throws Exception {
-        parameter.setCity(BuilderUtil.buildCityBoyarka());
-        parameter.setFree(false);
-
-        String expectedSql = "SELECT * FROM event e " +
-                "WHERE (e.event_date > NOW() AND city_id = #{city.id} AND price > 0) " +
-                "ORDER BY event_date LIMIT #{limit} OFFSET #{offset}";
-
-        String returnedSql = new EventSqlBuilder().getFilteredEvents(parameter).replace('\n', ' ');
-        assertEquals(expectedSql, returnedSql);
-    }
-
-    @Test
-    public void shouldBuiltSqlQueryByFewTechnologiesAndPagination() throws Exception {
-
-        List<Technology> testTechnologies = new ArrayList<>();
-        testTechnologies.add(BuilderUtil.buildTechnologyPhp());
-        testTechnologies.add(BuilderUtil.buildTechnologyAnt());
-        testTechnologies.add(BuilderUtil.buildTechnologySql());
-        parameter.setTechnologies(testTechnologies);
-
-        Iterator<Technology> iterator = testTechnologies.iterator();
-
-        String expectedSql = "SELECT * FROM event e JOIN event_technology et " +
-                "ON et.technology_id=" + iterator.next().getId() + " or et.technology_id=" + iterator.next().getId() +
-                " or et.technology_id=" + iterator.next().getId() + " WHERE (e.event_date > NOW() AND e.id=et.event_id) " +
-                "ORDER BY event_date LIMIT #{limit} OFFSET #{offset}";
-
-        String returnedSql = new EventSqlBuilder().getFilteredEvents(parameter).replace('\n', ' ');
-        assertEquals(expectedSql, returnedSql);
-    }
-
-    @Test
-    public void shouldBuildSqlQueryWithPaginationToGetEventsInRadius() throws Exception {
-        double testLatitude = 50.454605;
-        double testLongitude = 30.403965;
-        int testRadius = 5000;
-
-        parameter.setLatitude(testLatitude);
-        parameter.setLongitude(testLongitude);
-        parameter.setRadius(testRadius);
-
-        String expectedSql = "SELECT * FROM event e " +
-                "WHERE (e.event_date > NOW() " +
-                "AND ST_DWithin((point)::geography, ST_MakePoint(#{longitude},#{latitude})::geography, #{radius})) " +
-                "ORDER BY event_date LIMIT #{limit} OFFSET #{offset}";
-
-        String returnedSql = new EventSqlBuilder().getFilteredEvents(parameter).replace('\n', ' ');
-        assertEquals(expectedSql, returnedSql);
-    }
-
     @Test
     public void shouldBuildSqlQueryToAddFilterTechnologiesOfGivenFilter() throws Exception {
         Filter filter = BuilderUtil.buildFilterTest();
-        FilterSqlBuilder builder = new FilterSqlBuilder();
-
         String expectedSql = "INSERT INTO filter_technology (filter_id, technology_id) VALUES (-6, -1), (-6, -2)";
 
-        String returnedSql = builder.addFilterTechnology(filter);
+        String returnedSql = new FilterSqlBuilder().addFilterTechnology(filter);
 
         assertEquals(expectedSql, returnedSql);
     }
-
-    @Test
-    public void shouldBuildSqlQueryWithDateRangeWithRating() throws Exception {
-        Filter filter = BuilderUtil.buildTestFilter();
-        String expectedSql = "SELECT * FROM event e " +
-                "JOIN event_technology et ON et.technology_id=-1 or et.technology_id=-5 or et.technology_id=-8 " +
-                "LEFT OUTER JOIN  (SELECT event_id, COUNT(*) as visits FROM visit_log GROUP BY event_id) " +
-                    "AS visits ON e.id = visits.event_id " +
-                "WHERE (" +
-                    "e.event_date > NOW() " +
-                    "AND city_id = #{city.id} " +
-                    "AND (price IS NULL OR price = 0) " +
-                    "AND e.id=et.event_id " +
-                    "AND e.event_date < NOW() + (#{rangeInDays} || ' DAYS')::INTERVAL" +
-                ") " +
-                "ORDER BY event_date " +
-                "LIMIT #{limit}";
-        String returnedSql = new EventSqlBuilder().getFilteredEventsInDateRangeWithRating(filter)
-                .replace('\n', ' ');
-        assertEquals(expectedSql, returnedSql);
-    }
-
 }
